requirements.txt
selenium==4.28.1
beautifulsoup4==4.13.3
python-dotenv==1.0.1
Flask==2.2.5
Flask-Cors==5.0.0
redis==5.2.1
pytz==2024.2
gitpython
